[#774] moved resource labeling view function to hs_labels app
