+++ hs_labels/urls.py
@@ -1,0 +1,6 @@
+from django.conf.urls import patterns, url
+from hs_labels import views
+
+urlpatterns = patterns('',
+    url(r'^_internal/(?P<shortkey>[A-z0-9]+)/label-resource-action/$', views.resource_labeling_action),
+    )
