+++ hydroshare/urls.py
@@ -56,6 +56,7 @@ urlpatterns += patterns('',
     #url('^hs_party/', include('hs_party.urls')),
     url('^hs_metrics/', include('hs_metrics.urls')),
     url('^hsapi/', include('hs_model_program.urls')),
+    url('^hsapi/', include('hs_labels.urls')),
 )
 
 if settings.DEBUG is False:   #if DEBUG is True it will be served automatically
